feat: auto-assign shop to chest on create, require shift+click to break shop chest
Agent-Logs-Url: https://github.com/SLNE-Development/surf-shop/sessions/726e368b-1963-4485-9390-a6cab63f8334

+++ surf-shop-paper/src/main/kotlin/dev/slne/surf/shop/paper/chest/ShopChestListener.kt
@@ -71,6 +71,18 @@ object ShopChestListener : Listener {
             return
         }
 
+        if (!player.isSneaking) {
+            event.isCancelled = true
+            player.sendText {
+                appendErrorPrefix()
+                error("Du musst schleichen (Shift + Klick), um eine Shop Chest abzubauen.")
+            }
+            player.playSound(true) {
+                type(Sound.ENTITY_VILLAGER_NO)
+            }
+            return
+        }
+
         plugin.launch {
             staticShopChestService.deleteChest(chest.chestUuid)
